Update docs (#189)
* Update outdated links

* Update getting-started.md

- Capitalise Juju and MicroK8s where appropriate
- Update outdated commands
- Remove broken links in the "What you'll do" section
- Remove words for conciseness

* Update configure-restart-time-range.md

Add a "the"

* Update backup-and-restore-jenkins.md

- Capitalise Pebble where appropriate
- Update list items to contain the same grammatical structure.

* Update configure-installable-plugins.md

- Moved the code block higher in the section (I think it makes more sense for the `update-status` text to appear after the code block; please let me know if you disagree!)
- Fixed a fragmented sentence
- Other grammatical changes

* Update configure-jenkins-memory-usage.md

- Capitalise Juju where appropriate
- Change the "considerations" header to be one level lower than the "how to" header

* Update outdated command in get-admin-password.md

I removed the `--wait` flag because it now requires an argument. I'm not opposed to adding it back, but I would need guidance on an appropriate value to include.

* Change a header in integrate-with-external-agents.md

Made "Networking considerations" heading one level lower than the "how to" heading.

* Update redeploy.md

Removed the second section and moved text around for conciseness. Also updated the steps to have the same grammatical structure.

* fixed broken link

* Update resize-jenkins-storage.md

Fix broken links; small grammatical changes

* Update rotate-credentials.md

Updated an outdated command (removed the `--wait` flag, but not opposed to putting it back in); explicitly labeled the password as new

* Update external-access.md

Fixed a fragmented sentence; grammatical change in last sentence